// src/modules/inventory/enums/inventory-movement-type.enum.ts
export enum InventoryMovementType {
  RESTOCK = 'RESTOCK',
  SALE = 'SALE',
  RETURN = 'RETURN',
  ADJUSTMENT = 'ADJUSTMENT',
  RESERVATION = 'RESERVATION',
  RELEASE_RESERVATION = 'RELEASE_RESERVATION',
  DAMAGE = 'DAMAGE',
  THEFT = 'THEFT',
  TRANSFER_IN = 'TRANSFER_IN',
  TRANSFER_OUT = 'TRANSFER_OUT',
  EXPIRED = 'EXPIRED',
  QUALITY_CONTROL = 'QUALITY_CONTROL',
  MANUAL_CORRECTION = 'MANUAL_CORRECTION',
  SYSTEM_CORRECTION = 'SYSTEM_CORRECTION',
}